Fix Source Install Disk Detection in get_disk_list
The original check doesn't always work because PARTITIONS get labels, NOT disks. If you install with a Flash Drive and deploy to a machine with an NVMe drive, the Flash Drive is listed first because SCSIs are shown first by lsblk (At least they were in our scenario). When we run 'blkid "[Install Flash Drive Disk]" -s LABEL -o value' in this scenario, we get nothing returned so this conditional evaluates incorrectly. We should get the ${kiwi_install_volid} Partition Block Name, and search for whether ${disk_device} is contained within it as the Disk Device that corresponds to the Install Partition.

Ex:
     Flash Drive ${disk_device} '/dev/sda' is a substring of ${kiwi_install_disk_part} '/dev/sda1', which has a LABEL/${kiwi_install_volid} of 'INSTALL'. Therefore, we should ignore this ${disk_device}

Because Disk Device + Partition Block Names are standardized and additive for NVMes, SCSIs, and every other Type, this method of detection of the Install Device will work under every scenario. The Disk Name is always the first part of the Partition Name.