Fix Source Install Disk Detection in get_disk_list (#1982)
Fix Source Install Disk Detection in get_disk_list

The original check doesn't always work because PARTITIONS get labels, NOT disks. If you install with
a Flash Drive and deploy to a machine with an NVMe drive, the Flash Drive is listed first because SCSIs
are shown first by lsblk (At least they were in our scenario). When we run

blkid "[Install Flash Drive Disk]" -s LABEL -o value

in this scenario, we get nothing returned so this conditional evaluates incorrectly. We should get the
${kiwi_install_volid} Partition Block Name, and search for whether ${disk_device} is contained within it as
the Disk Device that corresponds to the Install Partition.