Prevent warning message on resize exception
On systems which are configured to run the oem resize at
every boot (default case) kiwi checks how much space is free
and if that fits the constraints configured as part of the
image description. If the constraints are not met a warning
message is displayed and the boot continuous without any
resize action happening.

This warning message however, always appears after the first
boot when the resize had happened and no rest space on disk
is present unless the disk geometry would have changed.
The situation of the reboot of the system without any disk
geometry change is the standard case and happens way more
often than the reboot with a disk geometry change.

Therefore the warning message displayed is not actually
a real warning and most often considered as an issue
when there is none. To relax this situation, this commit
only shows the warning message if the detected free space
on disk is greater than 2M, which is the condition under
which we assume an intentional (user made) disk geometry
change.

This Fixes #1958