Add global --kiwi-file option
When building with kiwi a search on the kiwi main config
file is made inside of the given --description directory.
The search looks up for the file config.xml or *.kiwi.
So far there was no opportunity to specify another name.
This commit adds an option in the global area named:

    --kiwi-file name

which will make kiwi to lookup this file inside of the
given --description directory and fail if it does not
exist. This Fixes #1973