Switch to stdlib json package
simplejson is a backport of the stdlib json package for older
python versions. It went into stdlib with 3.4, and it is being developed
there since then. the comment was that simplejson supports tuples, but
I see no difference with python 3.6+:

>>> myvar=(1,2,3)
>>> import json
>>> json.dumps(myvar)
'[1, 2, 3]'
>>> import simplejson
>>> simplejson.dumps(myvar)
'[1, 2, 3]'

On the other hand, json is simply faster than simplejson, and it is one
less dependency to worry about. here's a comparison (x86_64 AMD Renoir,
1 million invocations)

      json dumps 0.918306345003657 seconds
simplejson dumps 1.505740466993302 seconds
      json loads 0.793854812043719 seconds
simplejson loads 0.879414324008394 seconds