Add commandline option --bootinclude
If the kiwi builtin boot image(initrd) is used instead
of dracut via the initrd_system="kiwi" attribute, there is
an opportunity to tell kiwi to add packages listed
in the system image description also to the list of packages
used for building the root system from which the boot
image(initrd) is created. We call this feature a bootinclude.

As of now the user had to specify that a package should be
used for the system and the boot image as part of the image
description <package> definition like in the following
example:

    <package name="foo" bootinclude="true"/>

With this commit the information can also be provided on
the commandline like in the following example

    kiwi-ng system build ... --bootinclude foo

The option can be specified multiple times if more than
one package should be marked to be used for the boot image
prepare step.