Allow "uninitialized" as content of /etc/machine-id
According to machine-id(5), an empty file does not signal that this is the
first boot of the system. Instead, the file needs to be missing or contain
the string "uninitialized". A missing file does not work if the filesystem
is initially mounted read-only, so allow "uninitialized" as well, instead
of truncating it.

Improve the documentation of the method, e.g. dracut is not involved.