Ensure SELinux labels are set based on the policy
When running kiwi from a filesystem tree that has custom labels applied
(such as when using kiwi from within a container on an SELinux-enabled
host), the filesystem labeling doesn't correctly apply on some files
and folders with a warning about the location being customized by
the administrator. This causes all kinds of strange results with
the built images and makes them unbootable.

To resolve this, tell setfiles to forcibly set files and folders
with the default context from the installed policy.