Fixup inplace podman storage and container conf
Newer versions of podman requires runroot and graphroot
to be explicitly set in storage.conf.

Newer versions of podman no longer reads the engine.cgroups
setting on containers.conf and prints a 'Failed to decode the
keys [\"engine.cgroups\"]' warning message

This commit fixes storage.conf and containers.conf written
by kiwi if the setupContainerRuntime method is used in
scripts.