Fixed unconditional grub2 package requirement
The grub2 package does not exist on all distributions
as a name provider independent of the architecture.
On for example Debian and Ubuntu the packages are
handled differently and grub2 is only provided on
supported architectures. Thus the spec file should
set the grub2 requirement only if the distribution
provides it in any case