Fix btrfs volume mounting
If the second argument of os.path.join is an absolute directory, the
result would be that directory. The intention is to produce a
subdirectory of the mountpoint though. So pass a relative path.

Without the fix, kiwi would try to e.g mount the /var subvolume of
the image over the /var of the host, screwing everthing up of course
:-)