Stop copying /dev files statically into the OCI container
In containers (nspawn) where part of the /dev filesystem is bind-mounted
from outside system, kiwi fails to do the rsync (in creation of the
nodes).

There is no reason to actually copy whole tree inside so let's just
not do it (as it does not seem to be needed at all).