Disable gitlab CI pipeline
Unfortunately gitlab decided to allow only eligible
projects to use their CI pipeline. Being a member for
years seems not to count in this change process.
Therefore this commit disables the tests we run on
gitlab and I will come up with a simple github action
for the parts that we don't want to loose.