Swap from DM kpartx to util-linux partx for creating loop devices
The device-mapper (DM) based loop devices that kiwi has historically
generated creates issues when trying to run kiwi in confined build
environments. The DM tools prefer to have a tighter coupling between
userspace and kernel interfaces, and we cannot necessarily guarantee
that in the variety of build environments that kiwi can be run in.
In particular, Koji uses either nspawn containers or chroots through
Mock to run kiwi and that is where this fails.

However, we do not need to use DM for this purpose when util-linux
provides a perfectly serviceable alternative. This commit changes
kiwi's loop device setup to use partx(8) from util-linux instead.

As long as the appliance image being created doesn't use a disk
scheme that requires DM, it becomes entirely possible to produce
an image without needing DM at runtime at all.