Allow to specify fixed size for the root partition
So far the last partition (typically root) in the partition table
takes all the rest space of the partition table in the image file.
At deployment/boot time users had several options to let that
partition grow to a custom size. However, during build time of the
image it was not possible to specify a specific fixed size for the
root partition as we don't wanted to produce gaps of unpartitioned
space in the image file. It has turned out that there is hardware
available which requires a partition to be an exact multiple of
some blocksize. As kiwi supports size constraints for all other
partitions but not for root this commit now allows for it. The
oem-systemsize element is now also taken into account at build
time of the image if the oem-resize which would do that at
deployment/boot time is switched off like the following example
shows:

```xml
<oemconfig>
    <!-- set root partition to 2048MB -->
    <oem-systemsize>2048</oem-systemsize>
    <oem-resize>false</oem-resize>
</oemconfig>
```

This Fixes #2203