Add dracut-kiwi-verity dracut module package
When building an image with an embedded verity hash via the
embed_verity_metadata="true" type attribute it is required
to setup the device map at boot time accordingly. The dracut
module in this commit provides the needed tooling as a
POSIX C implementation which is able to read the kiwi embedded
verityhash from the offset of the partition. Kudos goes to
Isaac True <isaac.true@canonical.com>

Co-authored-by: Isaac True <isaac.true@canonical.com>
Signed-off-by: Isaac True <isaac.true@canonical.com>