Support EFI boot from squashfs
grub supports reading from squashfs. If no extra bootloader
installation is required this allows to boot from a squashfs
rootfs without extra boot partition. The prebuilt grub images
usually does not contain the squash4 module which is the
reason why kiwi creates the EFI module itself if the system
is fully immutable. Please note; the grub configuration
must be provided manually as part of the image description
in this case because grub-mkconfig at the time it has to be
called has no write permissions anymore.