Bring back create_parted_partitions method
This commit brings back the create_parted_partitions method
as it was before commit 1c48f33. Within the mentioned commit
the partitioner logic of the dracut modules dropped the use
of `parted` in favor of `sgdisk` and `sfdisk` depending on the
partition table type.

For s390 platforms it was used `fdasd`, however its seams that
nowadays support for DASD is better for `parted` compared with
`fdasd`. More specific, fdasd does not support resizing the
partition table and expanding partitions, but parted does.

For s390 parted should delete and recreate the partition
to be expanded, `resizepart` is not functional in that context.

According to SUSE s390 experts parted has superseded fdasd and
it should be the preferred tool for s390 whenever possible.

Within the same issue it has also been noted that `blkid` does not
properly detect dasd type, hence an heuristic approach has been
implemented based on the device name.

Fixes bsc#1209247

Signed-off-by: David Cassany <dcassany@suse.com>