apply grub-use-disk-password.patch;
This patch adds a new attribute to the <bootloader> element, called
use_disk_password (true/false). If set to true, the cryptomount command
in the EFI grub.cfg file will explicitly specify the LUKS password.

This is useful when shipping images with a LUKS root partition, allowing
them to come up on firstboot without the user having to enter a well-known
"secret".